Aggiunto formattatore per prezzi in Portata; modificato metodo allergeniChecker; dato un po' di colore a entrambi i metodi.

+++ src/Ristorante.java
@@ -9,8 +9,7 @@ public class Ristorante {
     //TODO ognugno di voi dovrà farsi la propria tabella quindi installiamo un db con lo stesso nome ma in locale sui vostri pc
     //TODO bisogna inserire la prenotazione usando una struttura dati mappa che mette in relazione cliente e tavolo che saranno due classi
     //nel main chiamare ristorante.prenota(cliente,tavolo) stampare la prenotazione e poi il menù in base al tipo di cliente(vegano,carn ecc ecc)
-    public Ristorante(String nome, String indirizzo, Byte numeroCoperti, Portata piattoDelGiorno) {
-        //TODO sistemare
+    public Ristorante(String nome, String indirizzo, Integer numeroCoperti, Portata piattoDelGiorno) {
         this.nome = nome;
         this.indirizzo = indirizzo;
         this.numeroCoperti = numeroCoperti;
@@ -33,11 +32,11 @@ public class Ristorante {
         this.indirizzo = indirizzo;
     }
 
-    public Byte getNumeroCoperti() {
+    public Integer getNumeroCoperti() {
         return numeroCoperti;
     }
 
-    public void setNumeroCoperti(Byte numeroCoperti) {
+    public void setNumeroCoperti(Integer numeroCoperti) {
         this.numeroCoperti = numeroCoperti;
     }
